fix: use riaSkey as primary key

+++ oracle-api/src/main/java/ca/bc/gov/oracleapi/repository/consep/GerminationTrayContentsRepository.java
@@ -2,6 +2,7 @@ package ca.bc.gov.oracleapi.repository.consep;
 
 import ca.bc.gov.oracleapi.entity.consep.GerminationTrayContentsEntity;
 import java.util.List;
+import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -10,7 +11,7 @@ import org.springframework.data.repository.query.Param;
  * Repository for germination tray contents view.
  */
 public interface GerminationTrayContentsRepository
-    extends JpaRepository<GerminationTrayContentsEntity, Integer> {
+    extends JpaRepository<GerminationTrayContentsEntity, BigDecimal> {
 
   @Query("""
       SELECT gtc
